[Update] Update Date from INT to LONG Datatype && Some bugs

+++ app/src/main/java/studentcompany/sportgest/domains/Game.java
@@ -63,7 +63,7 @@ public class Game extends DomainPojo {
         return date;
     }
 
-    public void setDate(int date) {
+    public void setDate(long date) {
         this.date = date;
     }
